+++ crates/lkjscript-ir/src/eval/instruction/constants.rs
@@ -14,7 +14,7 @@ impl Evaluator<'_> {
                 .static_bytes
                 .iter()
                 .position(|bytes| bytes.as_ref() == value)
-                .and_then(|index| u32::try_from(index).ok())
+                .and_then(|index| u64::try_from(index).ok())
                 .map(EvalValue::StaticBytes)
                 .ok_or_else(|| Flow::Trap("evaluator static bytes table mismatch".into())),
             Constant::Symbol(value) => self
